+++ admin/language/dutch/total/mollie_payment_fee.php
@@ -1,0 +1,17 @@
+<?php
+// Heading
+$_['heading_title']         = "Mollie Payment Fee";
+
+// Text
+$_['text_edit']               = "Bewerk Mollie";
+$_['text_extension']          = 'Extensions';
+$_['text_success']            = "Gelukt: de instellingen voor de module zijn aangepast!";
+$_['text_enable']             = "Activeren";
+$_['text_disable']            = "Deactiveren";
+
+// Entry
+$_['entry_status']	 		= "Status";
+$_['entry_sort_order']      = "Sorteervolgorde";
+
+// Error
+$_['error_permission']      = "Waarschuwing: U heeft geen toestemming om de module aan te passen.";