+++ catalog/language/nl-nl/extension/total/mollie_payment_fee.php
@@ -1,0 +1,2 @@
+<?php
+    include(__DIR__."/../../../dutch/total/mollie_payment_fee.php");